feat: refactor player and add interact event
