+++ src/level/plugin.rs
@@ -9,7 +9,9 @@ use crate::items::systems::toggle_on_system::toggle_on_enter_system;
 use crate::items::systems::update_is_activated_system::update_is_activated_system;
 use crate::level::load_level::load_level;
 use crate::level::unload_level::unload_level;
-use crate::player::player::{player_input_system, Player, PlayerPlugin};
+use crate::player::components::player::Player;
+use crate::player::plugin::PlayerPlugin;
+use crate::player::systems::player_input_system::player_input_system;
 use crate::player::{Ghost, GhostActions, GhostNewPositionEvent, PlayerNewPositionEvent};
 use crate::state::GameState;
 use crate::time::{elapsed_time_from_start_rewind_system, ElapsedTimeFromStartRewind, StartTime};
