+++ src/player/components/ghost.rs
