+++ src/player/mod.rs
@@ -1,18 +1,23 @@
 pub mod actions;
 pub mod events {
     pub mod ghost_new_position_event;
+    pub mod interact_event;
     pub mod new_position_event;
     pub mod player_new_position_event;
     pub mod rewind_event;
 }
-pub mod ghost;
-pub mod ghost_actions;
-pub mod interact;
+pub mod components {
+    pub mod ghost;
+    pub mod player;
+}
+pub mod systems {
+    pub mod ghost_actions_system;
+    pub mod on_player_rewind_system;
+    pub mod player_input_system;
+}
 pub mod move_direction;
-pub mod player;
-// exports
-
+pub mod plugin;
+pub use components::ghost::Ghost;
 pub use events::ghost_new_position_event::GhostNewPositionEvent;
 pub use events::player_new_position_event::PlayerNewPositionEvent;
-pub use ghost::Ghost;
-pub use ghost_actions::{ghost_actions_system, GhostActions};
+pub use systems::ghost_actions_system::{ghost_actions_system, GhostActions};
